+++ app/routes/components/Popover/index.tsx
@@ -36,7 +36,7 @@ export default function PopoverPage() {
         </DocLayout.Content>
       )}
       {tabsValue === "source" && (
-        <DocLayout.Content>
+        <DocLayout.Content variant="files">
           <DocLayout.Files files={sourceCode} />
         </DocLayout.Content>
       )}